fix: debugfs auto created by the docker daemon on macOS (#2005)
* chore:move docker test from util to cmd layer


* fix: docker cmd


* add: docker logs


---------
