+++ utils/utils.go
@@ -11,9 +11,7 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"runtime"
 	"runtime/debug"
 	"strconv"
 	"strings"
@@ -29,7 +27,6 @@ import (
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
-	"golang.org/x/term"
 )
 
 var WarningSign = "\U000026A0"
@@ -213,23 +210,6 @@ func CheckFileExists(path string) bool {
 
 var Version string
 
-type DockerConfigStruct struct {
-	DockerImage string
-	Envs        map[string]string
-}
-
-var DockerConfig = DockerConfigStruct{
-	DockerImage: "ghcr.io/keploy/keploy",
-}
-
-func GenerateDockerEnvs(config DockerConfigStruct) string {
-	var envs []string
-	for key, value := range config.Envs {
-		envs = append(envs, fmt.Sprintf("-e %s=%s", key, value))
-	}
-	return strings.Join(envs, " ")
-}
-
 func attachLogFileToSentry(logger *zap.Logger, logFilePath string) error {
 	file, err := os.Open(logFilePath)
 	if err != nil {
@@ -411,148 +391,6 @@ const (
 	Native        CmdType = "native"
 	Empty         CmdType = ""
 )
-
-func convertPathToUnixStyle(path string) string {
-	// Replace backslashes with forward slashes
-	unixPath := strings.Replace(path, "\\", "/", -1)
-	// Remove 'C:'
-	if len(unixPath) > 1 && unixPath[1] == ':' {
-		unixPath = unixPath[2:]
-	}
-	return unixPath
-}
-
-func getAlias(ctx context.Context, logger *zap.Logger) (string, error) {
-	// Get the name of the operating system.
-	osName := runtime.GOOS
-	//TODO: configure the hardcoded port mapping
-	img := DockerConfig.DockerImage + ":v" + Version
-	logger.Info("Starting keploy in docker with image", zap.String("image:", img))
-	envs := GenerateDockerEnvs(DockerConfig)
-	if envs != "" {
-		envs = envs + " "
-	}
-	var ttyFlag string
-
-	if term.IsTerminal(int(os.Stdin.Fd())) {
-		ttyFlag = " -it "
-	} else {
-		ttyFlag = ""
-	}
-
-	switch osName {
-	case "linux":
-		alias := "sudo docker container run --name keploy-v2 " + envs + "-e BINARY_TO_DOCKER=true -p 16789:16789 --privileged --pid=host" + ttyFlag + " -v " + os.Getenv("PWD") + ":" + os.Getenv("PWD") + " -w " + os.Getenv("PWD") + " -v /sys/fs/cgroup:/sys/fs/cgroup -v /sys/kernel/debug:/sys/kernel/debug -v /sys/fs/bpf:/sys/fs/bpf -v /var/run/docker.sock:/var/run/docker.sock -v " + os.Getenv("HOME") + "/.keploy-config:/root/.keploy-config -v " + os.Getenv("HOME") + "/.keploy:/root/.keploy --rm " + img
-		return alias, nil
-	case "windows":
-		// Get the current working directory
-		pwd, err := os.Getwd()
-		if err != nil {
-			LogError(logger, err, "failed to get the current working directory")
-		}
-		dpwd := convertPathToUnixStyle(pwd)
-		cmd := exec.CommandContext(ctx, "docker", "context", "ls", "--format", "{{.Name}}\t{{.Current}}")
-		out, err := cmd.Output()
-		if err != nil {
-			LogError(logger, err, "failed to get the current docker context")
-			return "", errors.New("failed to get alias")
-		}
-		dockerContext := strings.Split(strings.TrimSpace(string(out)), "\n")[0]
-		if len(dockerContext) == 0 {
-			LogError(logger, nil, "failed to get the current docker context")
-			return "", errors.New("failed to get alias")
-		}
-		dockerContext = strings.Split(dockerContext, "\n")[0]
-		if dockerContext == "colima" {
-			logger.Info("Starting keploy in docker with colima context, as that is the current context.")
-			alias := "docker container run --name keploy-v2 " + envs + "-e BINARY_TO_DOCKER=true -p 16789:16789 --privileged --pid=host" + ttyFlag + "-v " + pwd + ":" + dpwd + " -w " + dpwd + " -v /sys/fs/cgroup:/sys/fs/cgroup -v /sys/kernel/debug:/sys/kernel/debug -v /sys/fs/bpf:/sys/fs/bpf -v /var/run/docker.sock:/var/run/docker.sock -v " + os.Getenv("USERPROFILE") + "\\.keploy-config:/root/.keploy-config -v " + os.Getenv("USERPROFILE") + "\\.keploy:/root/.keploy --rm " + img
-			return alias, nil
-		}
-		// if default docker context is used
-		logger.Info("Starting keploy in docker with default context, as that is the current context.")
-		alias := "docker container run --name keploy-v2 " + envs + "-e BINARY_TO_DOCKER=true -p 16789:16789 --privileged --pid=host" + ttyFlag + "-v " + pwd + ":" + dpwd + " -w " + dpwd + " -v /sys/fs/cgroup:/sys/fs/cgroup -v debugfs:/sys/kernel/debug:rw -v /sys/fs/bpf:/sys/fs/bpf -v /var/run/docker.sock:/var/run/docker.sock -v " + os.Getenv("USERPROFILE") + "\\.keploy-config:/root/.keploy-config -v " + os.Getenv("USERPROFILE") + "\\.keploy:/root/.keploy --rm " + img
-		return alias, nil
-	case "darwin":
-		cmd := exec.CommandContext(ctx, "docker", "context", "ls", "--format", "{{.Name}}\t{{.Current}}")
-		out, err := cmd.Output()
-		if err != nil {
-			LogError(logger, err, "failed to get the current docker context")
-			return "", errors.New("failed to get alias")
-		}
-		dockerContext := strings.Split(strings.TrimSpace(string(out)), "\n")[0]
-		if len(dockerContext) == 0 {
-			LogError(logger, nil, "failed to get the current docker context")
-			return "", errors.New("failed to get alias")
-		}
-		dockerContext = strings.Split(dockerContext, "\n")[0]
-		if dockerContext == "colima" {
-			logger.Info("Starting keploy in docker with colima context, as that is the current context.")
-			alias := "docker container run --name keploy-v2 " + envs + "-e BINARY_TO_DOCKER=true -p 16789:16789 --privileged --pid=host" + ttyFlag + "-v " + os.Getenv("PWD") + ":" + os.Getenv("PWD") + " -w " + os.Getenv("PWD") + " -v /sys/fs/cgroup:/sys/fs/cgroup -v /sys/kernel/debug:/sys/kernel/debug -v /sys/fs/bpf:/sys/fs/bpf -v /var/run/docker.sock:/var/run/docker.sock -v " + os.Getenv("HOME") + "/.keploy-config:/root/.keploy-config -v " + os.Getenv("HOME") + "/.keploy:/root/.keploy --rm " + img
-			return alias, nil
-		}
-		// if default docker context is used
-		logger.Info("Starting keploy in docker with default context, as that is the current context.")
-		alias := "docker container run --name keploy-v2 " + envs + "-e BINARY_TO_DOCKER=true -p 16789:16789 --privileged --pid=host" + ttyFlag + "-v " + os.Getenv("PWD") + ":" + os.Getenv("PWD") + " -w " + os.Getenv("PWD") + " -v /sys/fs/cgroup:/sys/fs/cgroup -v debugfs:/sys/kernel/debug:rw -v /sys/fs/bpf:/sys/fs/bpf -v /var/run/docker.sock:/var/run/docker.sock -v " + os.Getenv("HOME") + "/.keploy-config:/root/.keploy-config -v " + os.Getenv("HOME") + "/.keploy:/root/.keploy --rm " + img
-		return alias, nil
-
-	}
-	return "", errors.New("failed to get alias")
-}
-
-func RunInDocker(ctx context.Context, logger *zap.Logger) error {
-	//Get the correct keploy alias.
-	keployAlias, err := getAlias(ctx, logger)
-	if err != nil {
-		return err
-	}
-	var quotedArgs []string
-
-	for _, arg := range os.Args[1:] {
-		quotedArgs = append(quotedArgs, strconv.Quote(arg))
-	}
-
-	var cmd *exec.Cmd
-
-	// Detect the operating system
-	if runtime.GOOS == "windows" {
-		var args []string
-		args = append(args, "/C")
-		args = append(args, strings.Split(keployAlias, " ")...)
-		args = append(args, os.Args[1:]...)
-		// Use cmd.exe /C for Windows
-		cmd = exec.CommandContext(
-			ctx,
-			"cmd.exe",
-			args...,
-		)
-	} else {
-		// Use sh -c for Unix-like systems
-		cmd = exec.CommandContext(
-			ctx,
-			"sh",
-			"-c",
-			keployAlias+" "+strings.Join(quotedArgs, " "),
-		)
-	}
-
-	cmd.Cancel = func() error {
-		return InterruptProcessTree(logger, cmd.Process.Pid, syscall.SIGINT)
-	}
-
-	cmd.Stdout = os.Stdout
-	cmd.Stdin = os.Stdin
-	cmd.Stderr = os.Stderr
-	logger.Debug("running the following command in docker", zap.String("command", cmd.String()))
-	err = cmd.Run()
-	if err != nil {
-		if ctx.Err() == context.Canceled {
-			return ctx.Err()
-		}
-		LogError(logger, err, "failed to start keploy in docker")
-		return err
-	}
-	return nil
-}
 
 // Keys returns an array containing the keys of the given map.
 func Keys(m map[string][]string) []string {
